chad/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chad"
description = 'CHat Agent in Django'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Kurtis Fields", email = "kurtisfields@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "loguru ==0.6.0",
  "Django ==4.2",
  "django-cors-headers ==3.14.0",
  "django-model-utils ==4.3.1",
  "channels ==4.0.0",
  "ariadne-django ==0.3.0",
  "PyJWT ==2.6.0"
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["uvicorn ==0.21.1", "black ==23.3.0"]
test = ["pytest ==7.2.1"]

[project.urls]
Documentation = "https://github.com/unknown/chad#readme"
Issues = "https://github.com/unknown/chad/issues"
Source = "https://github.com/unknown/chad"

[tool.hatch.version]
path = "chad/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "uvicorn",
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
develop = "uvicorn chad.asgi:application --reload"
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=chad --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "chad/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
